Updated android directory according to the latest RN build. Renamed spectrum to backpack as needed throughout.
