+++ android/app/_BUCK
@@ -35,12 +35,12 @@ android_library(
 
 android_build_config(
     name = "build_config",
-    package = "com.spectrum",
+    package = "com.backpack",
 )
 
 android_resource(
     name = "res",
-    package = "com.spectrum",
+    package = "com.backpack",
     res = "src/main/res",
 )
