+++ android/app/src/main/java/com/backpack/MainActivity.java
@@ -1,7 +1,8 @@
-package com.spectrum;
+package com.backpack;
+
+import android.os.Bundle;
 
 import com.facebook.react.ReactActivity;
-import android.os.Bundle;
 
 public class MainActivity extends ReactActivity {
 
@@ -11,9 +12,8 @@ public class MainActivity extends ReactActivity {
    */
   @Override
   protected String getMainComponentName() {
-    return "spectrum";
+    return "backpack";
   }
-
   @Override
   protected void onCreate(Bundle savedInstanceState) {
     super.onCreate(null);
