+++ android/settings.gradle
@@ -1,4 +1,4 @@
-rootProject.name = 'spectrum'
+rootProject.name = 'backpack'
 include ':react-native-fs'
 project(':react-native-fs').projectDir = new File(rootProject.projectDir, '../node_modules/react-native-fs/android')
 apply from: file("../node_modules/@react-native-community/cli-platform-android/native_modules.gradle"); applyNativeModulesSettingsGradle(settings)
